Move IO classes in their own package
