+++ src/main/java/com/github/aksc/InputOutput/CommandLineOutput.java
@@ -1,4 +1,6 @@
-package com.github.aksc;
+package com.github.aksc.InputOutput;
+
+import com.github.aksc.DerivationSystem;
 
 /**
  * Created by akselcakmak on 17/08/2018.
